Separate Service from VRUI
